requirements.txt
torchtext
zstandard
portalocker
wandb
torch
transformers
datasets
einops
seaborn
fancy_einsum
charset-normalizer
detoxify
cmapy
notebook
nnsight
transformer_lens
pytest
scikit-learn
pandas
levenshtein
concept-erasure
dropbox
python-dotenv
openai
trl
peft
bitsandbytes